fixed bug with auth and timing

+++ src/scripts/checkCookies.js
@@ -9,24 +9,29 @@ module.exports = function CheckCookies(obj,whoPASSED = 'dont know who'){
         return {value:objNew,isOK:result.isOK};
     }
 
-    if(obj.isOK === false){ 
-        if(result.isOK === true){
-            return ChangeState({...result})
-        }else{
-            return ChangeState({redirect:true});
+    if(!result.isOK){
+        return {isOK:result.isOK};
+    }else{
+        if(obj.isOK === false){ 
+            if(result.isOK === true){
+                return ChangeState({...result})
+            }else{
+                return ChangeState({redirect:true});
+            }
+    
+        }else if(obj.isOK === true){
+            
+            if(parseInt(result.timeCookies) <= 0){
+                deleteAllCookies();
+                return ChangeState({redirect:true});
+            }else{
+                return ChangeState({...result});
+            }    
+            
         }
-
-    }else if(obj.isOK === true){
-        
-        if(parseInt(result.timeCookies) <= 0){
-            deleteAllCookies();
-            return ChangeState({redirect:true});
-        }else{
-            return ChangeState({...result});
-        }
-
-        
     }
+
+    
 
     
 }
